+++ common/asmdef.h
@@ -1,5 +1,6 @@
 /* Copyright (c) 2002, 2005, 2006, 2007 Marek Michalkiewicz
    Copyright (c) 2006-2007 Dmitry Xmelkov
+   Copyright (c) 2025 Georg-Johann Lay
    All rights reserved.
 
    Redistribution and use in source and binary forms, with or without
@@ -37,11 +38,17 @@
    __zero_reg__
    XJMP
    XCALL
+   XIJMP
+   XICALL
+   do_prologue_saves
+   do_epilogue_restores
    XL, XH                     ; Via avr/common.h
    YL, YH                     ; Via avr/common.h
    ZL, ZH                     ; Via avr/common.h
 
    FUNCTION                   ; Just for #ifdef
+
+   PGMX_SECTION               ; formerly in libm/fplib/fp32def.h
 
    Provided asm .macros
    ====================
@@ -50,6 +57,8 @@
    ENTRY
    ENTRY_FLOAT
    ENDFUNC
+   OBJECT
+   ENDOBJ
 
    REGNO                      ; Helper for the X_ macros.
    X_movw
@@ -73,6 +82,12 @@
 
 #include "sectionname.h"
 
+
+/* Put constant tables in .progmemx in program memory, so they
+   don't nibble on the lower 64K PROGMEM.  */
+#define PGMX_SECTION(x)	.section .progmemx.gcc_##x, "a", "progbits"
+
+
 /* Macros in this header use local symbols with `.L__' prefix. */
 
 /* Historically, the _U() was intended to concatenate '_' prefix.
@@ -94,6 +109,54 @@
 # define XJMP	rjmp
 # define XCALL	rcall
 #endif
+
+#if defined(__AVR_HAVE_EIJMP_EICALL__)
+# define XIJMP  eijmp
+# define XICALL eicall
+#else
+# define XIJMP  ijmp
+# define XICALL icall
+#endif
+
+#ifndef __AVR_TINY__
+
+;; Prologue stuff
+;; do_prologue_saves  costs 12 bytes.
+;; __prologue_saves__ costs 56 bytes.
+.macro do_prologue_saves n_pushed n_frame=0
+    ldi r26, lo8(\n_frame)
+    ldi r27, hi8(\n_frame)
+    ldi r30, lo8(gs(.L_prologue_saves.\@))
+    ldi r31, hi8(gs(.L_prologue_saves.\@))
+    XJMP __prologue_saves__ + ((18 - (\n_pushed)) * 2)
+.L_prologue_saves.\@:
+.endm
+
+;; Epilogue stuff
+;; do_epilogue_restores  costs 12 bytes.
+;; __epilogue_restores__ costs 54 bytes.
+.macro do_epilogue_restores n_pushed n_frame=0
+    in      r28, SPL_IO_ADDR
+#ifdef SPH
+    in      r29, SPH_IO_ADDR
+.if \n_frame > 63
+    subi    r28, lo8(-\n_frame)
+    sbci    r29, hi8(-\n_frame)
+.elseif \n_frame > 0
+    adiw    r28, \n_frame
+.endif
+#else
+    clr     r29
+.if \n_frame > 0
+    subi    r28, lo8(-\n_frame)
+.endif
+#endif /* HAVE SPH */
+    ldi     r30, \n_pushed
+    XJMP __epilogue_restores__ + ((18 - (\n_pushed)) * 2)
+.endm
+
+#endif /* AVR_TINY */
+
 
 /* Macro FUNCTION is intended to start a function body without an entry.
    It is needed where an entry is at the middle of function, like in
@@ -120,6 +183,7 @@
 	.size	_U(\name), .L__END - .
 .endm
 #define	FUNCTION  FUNCTION	/* for CPP conditional expressions	*/
+
 
 /* Macro ENTRY is intended to start a function with entry point at the
    head of body. It is not needed to use a FUNCTION macro in this case.
@@ -160,6 +224,27 @@ _U(\lname):
 .macro	ENDFUNC
 .L__END:
 .endm
+
+
+/* Defines an object and sets .size, .type and the label.  Use like:
+   OBJECT [.global] my_data
+   ...
+   ENDOBJ my_data  */
+.macro OBJECT arg1 arg2
+    .ifc \arg2,
+        .type \arg1, "object"
+        \arg1:
+    .else
+        \arg1 \arg2
+        .type \arg2, "object"
+        \arg2:
+    .endif
+.endm
+
+.macro ENDOBJ name
+    .size \name, . - \name
+.endm
+
 
 /* Macro REGNO set the symbol name 'name' to integer which is a 'reg'
    number (0..31). If 'reg' is not a valid register, compilation error.
@@ -209,9 +294,6 @@ _U(\lname):
 /* Macro 'X_movw' extends enhanced movw instruction for classic chips.
  */
 .macro	X_movw	dst,src
-#if  __AVR_HAVE_MOVW__
-	movw	\dst,\src
-#else
   REGNO	.L__movw_dst, \dst
   .if	.L__movw_dst < 0
     .exitm
@@ -229,9 +311,14 @@ _U(\lname):
     .err	; Invalid src arg in X_movw macro.
     .exitm
   .endif
+  .if	.L__movw_dst != .L__movw_src
+#if  __AVR_HAVE_MOVW__
+	movw	\dst,\src
+#else
 	mov	.L__movw_dst, .L__movw_src
 	mov	.L__movw_dst + 1, .L__movw_src + 1
 #endif
+  .endif
 .endm
 
 /* Macro 'X_sbiw' extends SBIW instruction for AVR_TINY chips. */
@@ -356,51 +443,64 @@ _U(\lname):
    r0, r1 (__zero_reg__ - assumed to always contain 0), r30, r31.
 
    LPM_R0_ZPLUS_FINI is used atfer all reads to restore RAMPZ
-   in the case a restoration is required.
- */
+   in the case a restoration is required.  */
 
-	.macro	LPM_R0_ZPLUS_INIT hhi
+.macro	LPM_R0_ZPLUS_INIT hhi
 #if __AVR_ENHANCED__
   #if __AVR_HAVE_ELPM__
 	out	RAMPZ_IO_ADDR, \hhi
   #endif
 #endif
-	.endm
+.endm
 
-	.macro	LPM_R0_ZPLUS_NEXT hhi
+.macro	LPM_R0_ZPLUS_NEXT hhi, dst=r0
+  REGNO	.L__elpm_dst, \dst
+  .if .L__elpm_dst < 0
+    .exitm	; do not multiply errors
+  .endif
+
 #if __AVR_ENHANCED__
   #if __AVR_HAVE_ELPM__
-    /* ELPM with RAMPZ:Z post-increment, load RAMPZ only once */
-	elpm	r0, Z+
+	;; ELPM with RAMPZ:Z post-increment, load RAMPZ only once.
+	elpm	.L__elpm_dst, Z+
   #else
-    /* LPM with Z post-increment, max 64K, no RAMPZ (ATmega83/161/163/32) */
-	lpm	r0, Z+
+	;; LPM with Z+, max 64K, no RAMPZ (ATmega83/161/163/32).
+	lpm	.L__elpm_dst, Z+
   #endif
 #else
   #if __AVR_HAVE_ELPM__
-    /* ELPM without post-increment, load RAMPZ each time (ATmega103) */
+	;; ELPM without post-increment, load RAMPZ each time (ATmega103).
 	out	RAMPZ_IO_ADDR, \hhi
 	elpm
 	adiw	r30,1
 	adc	\hhi, __zero_reg__
+	.if .L__elpm_dst > 0
+		mov	.L__elpm_dst, r0
+	.endif
+  #elif defined (__AVR_TINY__)
+	;; AVRrc: Program is visible in the RAM address-space at offset 0x4000.
+	ld	.L__elpm_dst, Z+
   #else
-    /* LPM without post-increment, max 64K, no RAMPZ (AT90S*) */
+	;; LPM without post-increment, max 64K, no RAMPZ (AT90S*).
 	lpm
 	adiw	r30,1
+	.if .L__elpm_dst > 0
+		mov	.L__elpm_dst, r0
+	.endif
   #endif
 #endif
-	.endm
+.endm
 
 /* We only have to restore RAMPZ when the device has the RAMPD (sic!)
    register.  In that case, we have to restore RAMPZ to zero because
    RAMPZ is concatenated with Z to get 24-bit addresses for RAM access.
    When the device has no RAMPD, then the device has no RAMPZ, or RAMPZ
    is used exclusively with ELPM and need not to be restored.  */
-	.macro	LPM_R0_ZPLUS_FINI
+.macro	LPM_R0_ZPLUS_FINI
 #if defined(__AVR_HAVE_ELPM__) && defined(__AVR_HAVE_RAMPD__)
 	out	RAMPZ_IO_ADDR, __zero_reg__
 #endif
-	.endm
+.endm
 
 
 .macro TINY_WEAK_ALIAS new old