+++ tests/simulate/stdlib/strtoul.h
@@ -1,4 +1,5 @@
 /* Copyright (c) 2007  Dmitry Xmelkov
+   Copyright (c) 2025  Georg-Johann Lay
    All rights reserved.
 
    Redistribution and use in source and binary forms, with or without
@@ -32,8 +33,9 @@
 #ifndef	__AVR__
 # include <stdio.h>
 #endif
-#ifndef	EINVAL		/* Addition for errno.h in AVR-LibC-1.2.0	*/
-# define EINVAL	22	/* Invalid argument	*/
+
+#if __SIZEOF_LONG__ != 4
+#error expecting sizeof(long) == 4
 #endif
 
 static int
@@ -42,7 +44,9 @@ t_strtoul (const char *s, int base, unsigned long ret, int err, int len)
     char * endptr;
     
     errno = 0;
-    endptr = (char *)s - 1;		/* invalid value	*/
+    endptr = (char *)s;
+    __asm ("" : "+r" (endptr));
+    endptr--;		/* invalid value	*/
     if (strtoul (s, & endptr, base) != ret
 	|| errno != err
 #ifdef	__AVR__
@@ -52,7 +56,7 @@ t_strtoul (const char *s, int base, unsigned long ret, int err, int len)
 #endif
     {
 #ifndef	__AVR__
-	printf ("strtoul(\"%s\",,%d): %lu, errno: %d, len: %d\n",
+	printf ("strtoul(\"%s\",,%d): %lu, errno: %d, len: %zd\n",
 	    s, base, strtoul(s, & endptr, base), errno, endptr - s);
 #endif
 	return 1;